pyproject.toml
[project]
name = "youtube-ai-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
